fix(core): cast DSRN surprise error and softplus to float32 to prevent bf16 overflow
The surprise error (diff^2 clamped to 10.0) and softplus(surprise_lambda)
still ran in bf16, producing inf values in the backward when parameters
drifted to extreme values late in training (epoch ~0.97 of Stage 2).

Compute both in float32 then cast back, matching the gate sigmoid/tanh fix
from 86d052c.  All numerically sensitive ops in the DSRN kernel are now
float32-safe.

+++ echo_dsrn/modeling_echo.py
@@ -235,9 +235,11 @@ def dsrn_parallel_kernel_legacy(
 
     x_pred = model_block.linear_pred(h_shifted)
     diff = x - x_pred
-    error = torch.clamp(diff * diff, max=10.0).mean(dim=-1, keepdim=True)
+    error = torch.clamp((diff * diff).float(), max=10.0).to(x.dtype).mean(dim=-1, keepdim=True)
     # Constrain surprise_lambda strictly positive to guarantee error opens the memory gate
-    surprise_signal = error * torch.nn.functional.softplus(model_block.surprise_lambda)
+    surprise_signal = error * torch.nn.functional.softplus(model_block.surprise_lambda.float()).to(
+        x.dtype
+    )
 
     # Gates
     gate_logits = model_block.linear_gate(h_all) + surprise_signal
@@ -319,9 +321,11 @@ def dsrn_parallel_kernel_hybrid(
 
     x_pred = model_block.linear_pred(h_shifted)
     diff = x - x_pred
-    error = torch.clamp(diff * diff, max=10.0).mean(dim=-1, keepdim=True)
+    error = torch.clamp((diff * diff).float(), max=10.0).to(x.dtype).mean(dim=-1, keepdim=True)
     # Constrain surprise_lambda strictly positive to guarantee error opens the memory gate
-    surprise_signal = error * torch.nn.functional.softplus(model_block.surprise_lambda)
+    surprise_signal = error * torch.nn.functional.softplus(model_block.surprise_lambda.float()).to(
+        x.dtype
+    )
 
     gate_logits = model_block.linear_gate(h_all) + surprise_signal
     g_all = torch.sigmoid(gate_logits.float()).to(x.dtype)
